pmv-ads-2025-2-e4-proj-infra-t3-buymove-main/src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
motor==3.7.1
pydantic==2.9.2
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
